Move test implementation to inner classes
